Fixed saving icons with newest PyObjC
When extracting the data out of an image file, we need to get the actual bytes to write to disk,
this is done with the .bytes() call. On PyObjC >= 2.3 .bytes() returns a memoryview object instead
of the actual bytes, so .tobytes() needs to be called in order to fetch the actual image bytes.