Fix screen-sharing on OSX >= 10.10
For some unknown reason, the VNC viewer in OSX >= 10.10 needs to establish
2 TCP connections in order to complete the handshake. The first connection is
closed as soon as the VNC server sends the supported authentication mechanisms,
and the viewer starts a new connection, which succeeds this time.

Why? Go figure.

The fix consists of custom screen-sharing viewer and server handlers, which can
deal with the disconnects and act accordingly.