Avoid Cocoa autoreleased objects from being generated in non-GUI threads
It seems that assigning an attribute on an instance of any class that
inherits from NSObject triggers the creation of a __NSCFString that is
autoreleased. This may have something to do with the fact that NSObject
subclasses use __slots__ for their attributes which makes them properties
that execute unspecified Cocoa code inside PyObjc.

Nevertheless, avoid triggering them from non-GUI threads to avoid the
need of autorelease pools.